LSR-881 move WebContent class to Models group
